requirements.txt
smolagents[gradio]
python-dotenv
Pillow
litellm
ipykernel
cryptography
google-genai
ipywidgets